refactor(frontend): flatten new-acl subdirectory into acls
Move all files from acls/new-acl/ up to acls/ and update all import paths.
Eliminates the unnecessary nesting now that the legacy acl-list.tsx is removed.
