+++ frontend/src/components/pages/acls/acl-create-page.tsx
@@ -13,12 +13,12 @@ import { getRouteApi, useNavigate } from '@tanstack/react-router';
 
 const routeApi = getRouteApi('/security/acls/create');
 
-import { convertRulesToCreateACLRequests, handleResponses, type Rule } from 'components/pages/acls/new-acl/acl.model';
-import CreateACL from 'components/pages/acls/new-acl/create-acl';
-import { parsePrincipalFromParam, resolveAclSearchParams } from 'components/pages/acls/new-acl/principal-utils';
+import { convertRulesToCreateACLRequests, handleResponses, type Rule } from 'components/pages/acls/acl.model';
+import CreateACL from 'components/pages/acls/create-acl';
+import { parsePrincipalFromParam, resolveAclSearchParams } from 'components/pages/acls/principal-utils';
 
-import { useCreateAcls } from '../../../../react-query/api/acl';
-import { useSecurityBreadcrumbs } from '../../security/hooks/use-security-breadcrumbs';
+import { useCreateAcls } from '../../../react-query/api/acl';
+import { useSecurityBreadcrumbs } from '../security/hooks/use-security-breadcrumbs';
 
 const AclCreatePage = () => {
   const navigate = useNavigate({ from: '/security/acls/create' });
