+++ frontend/src/components/pages/acls/acl-detail-page.tsx
@@ -18,10 +18,10 @@ import { Pencil } from 'lucide-react';
 import { ACLDetails } from './acl-details';
 import { HostSelector } from './host-selector';
 import { parsePrincipalFromParam } from './principal-utils';
-import { useGetAclsByPrincipal } from '../../../../react-query/api/acl';
-import { Button } from '../../../redpanda-ui/components/button';
-import { Text } from '../../../redpanda-ui/components/typography';
-import { useSecurityBreadcrumbs } from '../../security/hooks/use-security-breadcrumbs';
+import { useGetAclsByPrincipal } from '../../../react-query/api/acl';
+import { Button } from '../../redpanda-ui/components/button';
+import { Text } from '../../redpanda-ui/components/typography';
+import { useSecurityBreadcrumbs } from '../security/hooks/use-security-breadcrumbs';
 
 const AclDetailPage = () => {
   const { aclName } = routeApi.useParams();
