+++ frontend/src/components/pages/acls/acl-details.tsx
