+++ frontend/src/components/pages/roles/role-detail-page.tsx
@@ -22,8 +22,8 @@ import { Button } from '../../redpanda-ui/components/button';
 import { Card, CardContent, CardHeader } from '../../redpanda-ui/components/card';
 import { Table, TableBody, TableCell, TableHead, TableHeader, TableRow } from '../../redpanda-ui/components/table';
 import { Text } from '../../redpanda-ui/components/typography';
-import type { AclDetail } from '../acls/new-acl/acl.model';
-import { ACLDetails } from '../acls/new-acl/acl-details';
+import type { AclDetail } from '../acls/acl.model';
+import { ACLDetails } from '../acls/acl-details';
 import { useSecurityBreadcrumbs } from '../security/hooks/use-security-breadcrumbs';
 
 type SecurityAclRulesTableProps = {
