+++ frontend/src/components/pages/roles/user-acls-card.tsx
@@ -16,8 +16,8 @@ import { useState } from 'react';
 import { Button } from '../../redpanda-ui/components/button';
 import { Card, CardAction, CardContent, CardHeader, CardTitle } from '../../redpanda-ui/components/card';
 import { Table, TableBody, TableCell, TableHead, TableHeader, TableRow } from '../../redpanda-ui/components/table';
-import { type AclDetail, getRuleDataTestId, parsePrincipal } from '../acls/new-acl/acl.model';
-import { OperationsBadge } from '../acls/new-acl/operations-badge';
+import { type AclDetail, getRuleDataTestId, parsePrincipal } from '../acls/acl.model';
+import { OperationsBadge } from '../acls/operations-badge';
 
 type UserAclsCardProps = {
   acls?: AclDetail[];
